requirements.txt
# Web Framework
Flask==3.0.2
Flask-Cors==4.0.0
Werkzeug==3.0.1
gunicorn==21.2.0

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.28
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
alembic==1.13.1

# Caching
redis==5.0.1
hiredis==2.3.2

# Web Scraping
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0

# AI/ML
groq==0.4.2

# Environment and Utils
python-dotenv==1.0.1
python-dateutil==2.8.2
pytz==2024.1

# Testing
pytest==8.0.2
pytest-cov==4.1.0

# Development
black==24.2.0
flake8==7.0.0
isort==5.13.2